REFACTOR: changed formatting on `CHMTableOfContents` class files.

+++ CHMTableOfContents.h
@@ -11,33 +11,41 @@
 
 
 @interface CHMTableOfContents : NSObject <NSOutlineViewDataSource> {
-	LinkItem *rootItems;
+	LinkItem			*rootItems;
 	
-	NSMutableArray *itemStack;
-	NSMutableArray *pageList;
-	LinkItem *curItem;
+	NSMutableArray		*itemStack;
+	NSMutableArray		*pageList;
+	LinkItem			*curItem;
+	
 }
+
 @property (readonly) LinkItem *rootItems;
 @property (readonly) NSArray *pageList;
 
-- (id)initWithData:(NSData *)data encodingName:(NSString*)encodingName;
+- (id)initWithData:(NSData *)data encodingName:(NSString *)encodingName;
 - (id)initWithTOC:(CHMTableOfContents*)toc filterByPredicate:(NSPredicate*)predicate;
+
 - (LinkItem *)curItem;
+
 - (LinkItem *)itemForPath:(NSString*)path withStack:(NSMutableArray*)stack;
+
 - (NSInteger)rootChildrenCount;
+
 - (void)sort;
+
 - (LinkItem*)getNextPage:(LinkItem*)item;
 - (LinkItem*)getPrevPage:(LinkItem*)item;
+
 @end
 
 
 @interface CHMSearchResult : CHMTableOfContents <NSOutlineViewDataSource> {
-	CHMTableOfContents* tableOfContent;
-	CHMTableOfContents* indexContent;
+	CHMTableOfContents *tableOfContent;
+	CHMTableOfContents *indexContent;
 }
 
 - (id)initwithTOC:(CHMTableOfContents*)toc withIndex:(CHMTableOfContents*)index;
-- (void)addPath:(NSString*)path score:(float)score;
+- (void)addPath:(NSString*)path score:(CGFloat)score;
 @end
